+++ src/setConfigQuickStep.ts
@@ -7,9 +7,16 @@ export async function setConfigQuickStep(
   project?: ZephyrProject
 ): Promise<string | undefined> {
 
+  let defaultName = 'setup';
+  if(project) {
+    if(project.configs.length > 0) {
+      defaultName = getNewConfigName(project.configs);
+    }
+  }
+
   const inputBox = vscode.window.createInputBox();
   inputBox.title = `Enter build configuration name`;
-  inputBox.value = context ? context.name : '';
+  inputBox.value = context && context.name.length > 0 ? context.name : defaultName;
   inputBox.prompt = 'Enter configuration name';
   inputBox.ignoreFocusOut = true;
 
@@ -31,7 +38,7 @@ export async function setConfigQuickStep(
         inputBox.validationMessage = `This "${input}" build configuration already exists`;
       }
     }
-});
+  });
 
   return new Promise((resolve) => {
     inputBox.onDidAccept(() => {
@@ -48,4 +55,27 @@ export async function setConfigQuickStep(
 
     inputBox.show();
   });
+}
+
+function getNewConfigName(configs: any[]): string {
+  const regex = /^setup(_(\d+))?$/;
+
+  const setupNumbers = configs
+    .map(config => {
+      const match = config.name.match(regex);
+      if (match && match[2]) {
+        return parseInt(match[2], 10);
+      } else if (config.name === 'setup') {
+        return 1;
+      }
+      return null;
+    })
+    .filter(num => num !== null) as number[];
+
+  if (setupNumbers.length === 0) {
+    return 'setup';
+  }
+
+  const latestSetupNumber = Math.max(...setupNumbers);
+  return `setup_${latestSetupNumber + 1}`;
 }